+++ src/main/java/dev/kc/pulse/mule/MuleAutoConfiguration.java
@@ -17,6 +17,7 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.boot.health.autoconfigure.contributor.ConditionalOnEnabledHealthIndicator;
 import org.springframework.boot.health.contributor.CompositeHealthContributor;
 import org.springframework.boot.health.contributor.HealthContributor;
 import org.springframework.boot.health.contributor.HealthIndicator;
@@ -28,6 +29,7 @@ import org.springframework.context.annotation.Bean;
 @AutoConfiguration(after = PulseAutoConfiguration.class)
 @ConditionalOnClass(HealthIndicator.class)
 @ConditionalOnProperty(prefix = "pulse.mule", name = "enabled", havingValue = "true")
+@ConditionalOnEnabledHealthIndicator("mule")
 @EnableConfigurationProperties(MuleProperties.class)
 public class MuleAutoConfiguration {