+++ src/main/java/dev/kc/pulse/oauth2/OAuth2AutoConfiguration.java
@@ -21,6 +21,7 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.boot.health.autoconfigure.contributor.ConditionalOnEnabledHealthIndicator;
 import org.springframework.boot.health.contributor.CompositeHealthContributor;
 import org.springframework.boot.health.contributor.HealthContributor;
 import org.springframework.boot.health.contributor.HealthIndicator;
@@ -33,6 +34,7 @@ import org.springframework.security.oauth2.client.registration.ClientRegistratio
 @AutoConfiguration(after = PulseAutoConfiguration.class)
 @ConditionalOnClass({ HealthIndicator.class, ObjectMapper.class, ClientRegistrationRepository.class })
 @ConditionalOnProperty(prefix = "pulse.oauth2", name = "enabled", havingValue = "true")
+@ConditionalOnEnabledHealthIndicator("oauth2")
 @EnableConfigurationProperties(OAuth2Properties.class)
 public class OAuth2AutoConfiguration {